+++ packages/mdx-components/.eslintrc.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  root: true,
+  extends: ['@it-incubator/eslint-config'],
+}