+++ packages/mdx-components/src/index.ts
@@ -1,0 +1,8 @@
+import '@fontsource/roboto'
+import './styles/index.scss'
+import { clsx } from 'clsx'
+export * from './assets/icons'
+export type { IconProps } from './assets/icons/IconWrapper'
+
+export { clsx }
+export * from './components'