+++ packages/mdx-components/storybook-utils/components/previews/value-preview/index.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react'
+
+export const ValuePreview = ({ children }: { children: ReactNode }) => {
+  return <div style={{ color: 'white', mixBlendMode: 'difference' }}>{children}</div>
+}